pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devdash"
version = "0.3.0"
description = "A terminal dashboard for developers - Git status, system stats, tasks, and Pomodoro timer in one screen"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["terminal", "dashboard", "git", "productivity", "pomodoro", "tui"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Terminals",
]

dependencies = [
    "textual>=0.47.0",
    "psutil>=5.9.0",
    "gitpython>=3.1.0",
    "tomli>=2.0.0;python_version<'3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[project.scripts]
devdash = "devdash.main:cli"

[project.urls]
Homepage = "https://github.com/RainMaker033/devdash"
Repository = "https://github.com/RainMaker033/devdash"
Issues = "https://github.com/RainMaker033/devdash/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
devdash = ["*.css", "*.tcss"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
